기본 필요기능들은 전부 구현 완료, 리팩토링과 기존 기능들 견고하게 시작
